refactor: use createItem hook to RankingSection
