+++ README.md
@@ -26,9 +26,9 @@
   - [x] hooks를 사용하여 필터 기능 구현 (ex. 전체, 여성이, 남성이, 청소년이 / 받고 싶어한, 많이 선물한, 위시로 받은)
   - [x] 상품 목록을 처음에는 6개만 보여지게 함.
   - [x] 더보기를 누르는 경우 상품 목록을 더 보여주도록 함. (접기 버튼을 누르면 다시 6개만 보여지게 됨.)
-- [ ] Theme 페이지 ( `/theme/:themeKey` )
-  - [ ] 재사용성을 고려하여 Header 섹션 구현 (themeKey에 따라 label, title, description, backgroundColor가 달라짐)
-  - [ ] 상품 목록 섹션 구현
+- [x] Theme 페이지 ( `/theme/:themeKey` )
+  - [x] 재사용성을 고려하여 Header 섹션 구현 (themeKey에 따라 label, title, description, backgroundColor가 달라짐)
+  - [x] 상품 목록 섹션 구현
 - [ ] 로그인 페이지( `/login` )
   - [ ] ID와 PW를 입력하면 로그인이 되도록 구현(ID와 PW는 아무 값을 입력해도 통과되도록 함.)
 - [ ] 나의 페이지(`/my-account`) 구현
